requirements.txt
aiofiles
aiohttp
argon2-cffi
beautifulsoup4
colorama
dashscope
fastapi
pymupdf
gradio
gradio_modal
latex2mathml
Markdown
numpy
pandas
pydantic
pymdown-extensions
PyPDF2
python_docx
PyYAML
rarfile
Requests
rich
rjsmin
starlette
tiktoken
uvicorn
websockets
zhipuai



